3.0.1.E.0.88

+++ kernel/arch/mn10300/mm/init.c
@@ -118,7 +118,8 @@ void __init mem_init(void)
 	       reservedpages << (PAGE_SHIFT - 10),
 	       datasize >> 10,
 	       initsize >> 10,
-	       totalhigh_pages << (PAGE_SHIFT - 10));
+	       (unsigned long) (totalhigh_pages << (PAGE_SHIFT - 10))
+	       );
 }
 
 /*
